Todo corregido

+++ BancApp/src/main/java/es/proyectotaw/banca/bancapp/dao/CambDivisaEntityRepository.java
@@ -1,6 +1,7 @@
 package es.proyectotaw.banca.bancapp.dao;
 
 import es.proyectotaw.banca.bancapp.entity.CambDivisaEntity;
+import es.proyectotaw.banca.bancapp.entity.TransferenciaEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.web.bind.annotation.RequestParam;
@@ -10,7 +11,7 @@ import java.util.List;
 public interface CambDivisaEntityRepository extends JpaRepository<CambDivisaEntity, Integer> {
 
     @Query("select c from CambDivisaEntity c where c.cantidad >= :cantidad")
-    List<CambDivisaEntity> filtrarPorCantidad(@RequestParam("cantidad") float cantidad);
+    List<CambDivisaEntity> filtrarPorCantidad(@RequestParam("cantidad") Double cantidad);
 
     @Query("select c from CambDivisaEntity c where c.operacionByOperacion.idOperacion = :id")
     CambDivisaEntity findByOperation(@RequestParam("id") int id);
