+++ BancApp/src/main/java/es/proyectotaw/banca/bancapp/dao/TransferenciaEntityRepository.java
@@ -11,11 +11,12 @@ import java.util.List;
 public interface TransferenciaEntityRepository extends JpaRepository<TransferenciaEntity, Integer> {
 
     @Query("select t from TransferenciaEntity t where t.cantidad >= :cantidad")
-    List<TransferenciaEntity> filtrarPorCantidad(@RequestParam("cantidad") float cantidad);
+    List<TransferenciaEntity> filtrarPorCantidad(@RequestParam("cantidad") double cantidad);
 
     @Query("select t from TransferenciaEntity t where t.ibanDestino in :sospechosas")
     List<TransferenciaEntity> filtrarPorDestinoSospechoso(@RequestParam("sospechosas") List<String> sospechosas);
 
     @Query("select t from TransferenciaEntity t where t.operacionByOperacion.idOperacion = :id")
     TransferenciaEntity findByOperation(@RequestParam("id") int id);
+
 }
